Moved things around for better organization
